Repair `gulp self` range of tasks.
This commit fixes up the various `gulp self` tasks. The path and file references have been moved out of `config.js` as a result of the change to allowing the Gulp tasks to be run from outside this package. 

References were incorrect and couldn't be corrected without hard coding. It was simpler, and makes more sense, to use `path` to resolve a relative path to the files which require linting.