Improve how style file headers are generated
All of the possible style file header pieces of data are now whitelisted against a `config.theme` key. Now, the theme can set the key with the value it wants - either hard-coded string, or pulled from `pkg`. For instance, a theme might have the following in its `Gulpfile.js` (which chooses not to set status or domain path for instance), which uses some top-level `package.json` keys, some custom ones from keys under the `theme` key, and a hard-coded string:

```
toolkit.extendConfig( {
		theme: {
		name: pkg.theme.name,
		themeuri: pkg.homepage,
		author: pkg.author.name,
		authoruri: pkg.author.url,
		description: pkg.theme.description,
		version: pkg.version,
		license: pkg.license,
		licenseuri: pkg.theme.licenseuri,
		textdomain: pkg.name,
		tags: pkg.theme.tags,
		template: 'genesis'
    }
} );
```

The default values in the toolkit's `config.js` are removed, since they would only pull in values from the toolkit's `package.json` if the theme didn't overwrite them anyway.

Note, that since Author URI and License URI are now supported, the config.theme key for the Theme URI is now `themeuri` instead of `homepage`. Consuming themes are still able to set this to use the `homepage` key in `package.json`, as per the above example.

Fixes #19