Drop Bower support
We can simplify the install process though, but letting themes add their dependencies to their `package.json`, and therefore letting them have those dependencies when they do the `yarn install` / `npm install` (when they'll pull in `gulp-wp-toolkit` as well).

This allows us to remove a fair bit of gulp-wp-toolkit, and requires one less tool to be installed on the theme developer's machine to be able to build a theme.

The two packages dropped already gave the following warnings anyway:

> warning bower-files@3.14.1: This module is no longer being maintained. Please reach out to package author if you wish to take over maintenance. For alternatives, use wiredep, or migrate to something like webpack
> warning gulp-bower > bower@1.8.0: ..psst! While Bower is maintained, we recommend Yarn and Webpack for *new* front-end projects! Yarn's advantage is security and reliability, and Webpack's is support for both CommonJS and AMD projects. Currently there's no migration path, but please help to create it: https://github.com/bower/bower/issues/2467