Add scripts to lint this package
The `gulp self` tasks couldn't be run without a Gulpfile present, and a local dependency on Gulp itself.

Instead, we move to use npm scripts, and the individual tools themselves (unwrapped from gulp).

There are three self-linting tasks:

- `jsonlint` - which validates some JSON files. More files can be added as needed.
- `esvalidate` - which ensures JavaScript files are valid. The output doesn't appear to show the file name, and the suggestion at https://github.com/duereg/esvalidate/issues/3 doesn't appear to work for me.
-  `eslint` - checks for good JavaScript coding practices and formatting. Rather than sticking with `eslint-config-wordpress` (which checks JS which will run in the browser), this uses the `eslint-config-gulp` config.

These can be run with, for instance, `npm run eslint`. All of the self-linting tasks can be in one go with `npm run lint`. `npm test` exists as a convenience too.

The eslintConfig in `package.json` avoids the need to have a separate `.eslintrc` file. We extend the `eslint-config-gulp` config, and specify we're using ES6.

The rules section is needed to avoid false positives regarding a change of rule name. These rules can be fixed as and when https://github.com/gulpjs/eslint-config-gulp/pull/13 gets merged.

See #25.