Fix `gulp bump` task.
As outlined in #87, there is an issue with the `gulp bump` task not
respecting parameters passed to it. I've gone with your suggested code,
@timothyjensen. As part of the updates I've found I can no longer pass
`--minor` as it's a Node Flag. That being the case, I've shortened the
arguments to the following:

`gulp bump` - a patch release, ie 1.0.0 to 1.0.2
`gulp bump --mi` - a minor release, ie 1.0.0 to 1.1.0
`gulp bump --ma` - a major release, ie 1.0.0 to 2.0.0

Fixes #87