Merge in develop branch. (#105)
* Update `.csscomb.json` to more WPCS friendly ruleset.

* Allow passing options to `cssnano` package.

As outlined in #104, CSSNano alters z-index attributes in the
stylesheet which causes issues when you're trying to override rules from
plugins or other CSS files.

See #104

* Remove style guide task.

* Fix `gulp bump` task.

As outlined in #87, there is an issue with the `gulp bump` task not
respecting parameters passed to it. I've gone with your suggested code,
@timothyjensen. As part of the updates I've found I can no longer pass
`--minor` as it's a Node Flag. That being the case, I've shortened the
arguments to the following:

`gulp bump` - a patch release, ie 1.0.0 to 1.0.2
`gulp bump --mi` - a minor release, ie 1.0.0 to 1.1.0
`gulp bump --ma` - a major release, ie 1.0.0 to 2.0.0

Fixes #87

* Update version number ahead of merge.